refactor: refactors based on review
