+++ apps/velo-external-db/test/storage/schema_provider.spec.ts
@@ -2,7 +2,7 @@ import Chance = require('chance')
 import { errors, SystemFields } from '@wix-velo/velo-external-db-commons'
 import { SchemaOperations } from '@wix-velo/velo-external-db-types'
 import { Uninitialized, gen, testIfSupportedOperationsIncludes } from '@wix-velo/test-commons'
-import { env, dbTeardown, setupDb, currentDbImplementationName, supportedOperations, currentDbCapabilities } from '../resources/provider_resources'
+import { env, dbTeardown, setupDb, currentDbImplementationName, supportedOperations } from '../resources/provider_resources'
 import { defaultFields, collectionWithFields, defaultCollection, hasSameSchemaFieldsLike } from '../drivers/schema_provider_matchers'
 const chance = new Chance()
 const { CollectionDoesNotExists, FieldAlreadyExists, CannotModifySystemField, FieldDoesNotExist } = errors
@@ -49,10 +49,9 @@ describe(`Schema API: ${currentDbImplementationName()}`, () => {
     })
 
     test('create collection with default columns', async() => {
-        const capabilities = currentDbCapabilities()
         await env.schemaProvider.create(ctx.collectionName)
 
-        await expect( env.schemaProvider.describeCollection(ctx.collectionName) ).resolves.toEqual(defaultCollection(ctx.collectionName, capabilities))
+        await expect( env.schemaProvider.describeCollection(ctx.collectionName) ).resolves.toEqual(defaultCollection(ctx.collectionName, env.capabilities))
     })
 
     test('drop collection', async() => {
@@ -64,17 +63,15 @@ describe(`Schema API: ${currentDbImplementationName()}`, () => {
     })
 
     test('collection name and variables are case sensitive', async() => {
-        const capabilities = currentDbCapabilities()
         await env.schemaProvider.create(ctx.collectionName.toUpperCase())
 
-        await expect( env.schemaProvider.describeCollection(ctx.collectionName.toUpperCase()) ).resolves.toEqual(defaultCollection(ctx.collectionName.toUpperCase(), capabilities))
+        await expect( env.schemaProvider.describeCollection(ctx.collectionName.toUpperCase()) ).resolves.toEqual(defaultCollection(ctx.collectionName.toUpperCase(), env.capabilities))
     })
 
     test('retrieve collection data by collection name', async() => {
-        const capabilities = currentDbCapabilities()
         await env.schemaProvider.create(ctx.collectionName)
 
-        await expect( env.schemaProvider.describeCollection(ctx.collectionName) ).resolves.toEqual(defaultCollection(ctx.collectionName, capabilities))
+        await expect( env.schemaProvider.describeCollection(ctx.collectionName) ).resolves.toEqual(defaultCollection(ctx.collectionName, env.capabilities))
     })
 
     test('create collection twice will do nothing', async() => {
@@ -88,10 +85,9 @@ describe(`Schema API: ${currentDbImplementationName()}`, () => {
     })
 
     test('add column on a an existing collection', async() => {
-        const capabilities = currentDbCapabilities()
         await env.schemaProvider.create(ctx.collectionName, [])
         await env.schemaProvider.addColumn(ctx.collectionName, { name: ctx.columnName, type: 'datetime', subtype: 'timestamp' })
-        await expect( env.schemaProvider.describeCollection(ctx.collectionName) ).resolves.toEqual(collectionWithFields(ctx.collectionName, [{ field: ctx.columnName, type: 'datetime' }], capabilities))
+        await expect( env.schemaProvider.describeCollection(ctx.collectionName) ).resolves.toEqual(collectionWithFields(ctx.collectionName, [{ field: ctx.columnName, type: 'datetime' }], env.capabilities))
     })
 
     test('add duplicate column will fail', async() => {
