+++ libs/velo-external-db-types/src/index.ts
@@ -1,4 +1,9 @@
-import { CollectionCapabilities, ColumnCapabilities } from './collection_types'
+import { 
+    ResponseField,
+    SchemaOperations,
+    ISchemaProvider
+} from './collection_types'
+
 export * from './collection_types'
 
 export enum AdapterOperator { //in velo-external-db-core
@@ -17,31 +22,6 @@ export enum AdapterOperator { //in velo-external-db-core
     not = 'not',
     urlized = 'urlized',
     matches = 'matches'
-}
-
-export enum SchemaOperations {
-    List = 'list',
-    ListHeaders = 'listHeaders',
-    Create = 'createCollection',
-    Drop = 'dropCollection',
-    AddColumn = 'addColumn',
-    RemoveColumn = 'removeColumn',
-    ChangeColumnType = 'changeColumnType',
-    Describe = 'describeCollection',
-    FindWithSort = 'findWithSort',
-    Aggregate = 'aggregate',
-    BulkDelete = 'bulkDelete',
-    Truncate = 'truncate',
-    UpdateImmediately = 'updateImmediately',
-    DeleteImmediately = 'deleteImmediately',
-    StartWithCaseSensitive = 'startWithCaseSensitive',
-    StartWithCaseInsensitive = 'startWithCaseInsensitive',
-    Projection = 'projection',
-    FindObject = 'findObject',
-    Matches = 'matches',
-    NotOperator = 'not',
-    IncludeOperator = 'include',
-    FilterByEveryField = 'filterByEveryField',
 }
 
 export type FieldWithQueryOperators = ResponseField & { queryOperators: string[] }
@@ -128,42 +108,28 @@ export interface IDataProvider {
     aggregate?(collectionName: string, filter: AdapterFilter, aggregation: AdapterAggregation, sort?: Sort[], skip?: number, limit?: number ): Promise<Item[]>;
 }
 
-export type Table = { 
-    id: string,
-    fields: ResponseField[]
-    capabilities?:  CollectionCapabilities
-} 
+// export type Table = { 
+//     id: string,
+//     fields: ResponseField[]
+//     capabilities?:  CollectionCapabilities
+// } 
 
-export type FieldAttributes = {
-    type: string,
-    subtype?: string,
-    precision?: number | string,
-    isPrimary?: boolean,
-}
+// export type FieldAttributes = {
+//     type: string,
+//     subtype?: string,
+//     precision?: number | string,
+//     isPrimary?: boolean,
+// }
 
-export type InputField = FieldAttributes & { name: string }
+// export type InputField = FieldAttributes & { name: string }
 
-export type ResponseField = FieldAttributes & { 
-    field: string
-    capabilities?: {
-        sortable: boolean
-        columnQueryOperators: string[]
-    }
-}
-export interface ISchemaProvider {
-    list(): Promise<Table[]>
-    listHeaders(): Promise<string[]>
-    supportedOperations(): SchemaOperations[]
-    create(collectionName: string, columns?: InputField[]): Promise<void>
-    addColumn(collectionName: string, column: InputField): Promise<void>
-    removeColumn(collectionName: string, columnName: string): Promise<void>
-    changeColumnType?(collectionName: string, column: InputField): Promise<void>
-    describeCollection(collectionName: string): Promise<ResponseField[]> | Promise<Table>
-    drop(collectionName: string): Promise<void>
-    translateDbTypes?(column: InputField | ResponseField | string): ResponseField | string
-    columnCapabilitiesFor?(columnType: string): ColumnCapabilities
-    capabilities?(): CollectionCapabilities
-}
+// export type ResponseField = FieldAttributes & { 
+//     field: string
+//     capabilities?: {
+//         sortable: boolean
+//         columnQueryOperators: string[]
+//     }
+// }
 
 export interface IBaseHttpError extends Error {}
